Need to get feedback, but basic work of extracting table is done
