+++ web-registry/src/components/organisms/CreditBatches.tsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import { makeStyles, useTheme } from '@mui/styles';
+import { makeStyles } from '@mui/styles';
+import Box from '@mui/material/Box';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableHead from '@mui/material/TableHead';
@@ -12,6 +13,7 @@ import {
   StyledTableContainer,
   StyledTableCell,
   StyledTableRow,
+  getTablePaginationPadding,
   // StyledTableSortLabel,
 } from 'web-components/lib/components/table';
 import { useTablePagination } from 'web-components/lib/components/table/useTablePagination';
@@ -100,7 +102,6 @@ const ROWS_PER_PAGE_OPTIONS = [5, 10];
 
 const CreditBatches: React.FC = () => {
   const styles = useStyles();
-  const theme = useTheme();
   const [batches, setBatches] = useState<any[]>([]);
   const [order /* , setOrder */] = useState<Order>('desc');
   const [orderBy /* , setOrderBy */] = useState<string>('start_date');
@@ -167,14 +168,6 @@ const CreditBatches: React.FC = () => {
     return num > 0 ? Math.floor(num).toLocaleString() : '-';
   };
 
-  const getPaddingBottom = (
-    countTotal: number,
-    rowsPerPage: number,
-    countPage: number,
-  ): number => {
-    return countTotal <= rowsPerPage ? 0 : rowsPerPage - countPage;
-  };
-
   return ledgerRestUri && batches.length > 0 ? (
     <Section
       classes={{ root: styles.section, title: styles.title }}
@@ -182,17 +175,16 @@ const CreditBatches: React.FC = () => {
       titleVariant="h2"
     >
       <StyledTableContainer className={styles.tableBorder}>
-        <div
-          style={{
+        <Box
+          sx={{
             width: '100%',
             overflow: 'auto',
-            paddingBottom: `${theme.spacing(
-              getPaddingBottom(
+            paddingBottom:
+              getTablePaginationPadding(
                 paginationProps.count,
                 paginationProps.rowsPerPage,
                 batches.length,
               ) * 25,
-            )}`,
           }}
         >
           <Table aria-label="credit batch table" stickyHeader>
@@ -268,7 +260,7 @@ const CreditBatches: React.FC = () => {
               )}
             </TableBody>
           </Table>
-        </div>
+        </Box>
         <Table>
           <TableFooter>
             <TableRow>
