+++ web-registry/src/components/organisms/index.ts
@@ -2,6 +2,7 @@ export * from './AdditionalityForm';
 export * from './AdminNav';
 export * from './AppFooter';
 export * from './BasicInfoForm';
+export * from './BuyCreditsModal';
 export * from './ConfirmationModal';
 export * from './CreditBatches';
 export * from './CreditClassCards';
@@ -9,6 +10,7 @@ export * from './CreditClassConnectSection';
 export * from './CreditClassOverviewSection';
 export * from './CreditsPurchaseForm';
 export * from './Documentation';
+export * from './EcocreditsTable';
 export * from './EntityDisplayForm';
 export * from './ImpactSection';
 export * from './IncludesGrasslandsForm';
@@ -33,4 +35,3 @@ export * from './StepCardsWithDescription';
 export * from './StepsSection';
 export * from './StoryForm';
 export * from './UserCredits';
-export * from './BuyCreditsModal';
