+++ web-registry/src/pages/index.ts
@@ -2,26 +2,27 @@ export * from './Additionality';
 export * from './Admin';
 export * from './BasicInfo';
 export * from './BuyerCreate';
+export * from './BuyerCreditsTransfer';
 export * from './Buyers';
 export * from './Certificate';
 export * from './ChooseCreditClass';
 export * from './CreateCreditClass';
-export * from './BuyerCreditsTransfer';
 export * from './CreateMethodology';
 export * from './CreditClassDetails';
 export * from './CreditsIssue';
 export * from './CreditsRetire';
 export * from './CreditsTransfer';
-export * from './Ecocredits';
+export * from './EcocreditsForAccount';
 export * from './Eligibility';
 export * from './EntityDisplay';
 export * from './GettingStarted';
 export * from './Home';
 export * from './LandStewards';
 export * from './Media';
-export * from './MethodologyDetails';
 export * from './MethodologiesList';
+export * from './MethodologyDetails';
 export * from './MethodologyReviewProcess';
+export * from './MyEcocredits';
 export * from './NotFound';
 export * from './OrganizationProfile';
 export * from './PostPurchase';
